requirements.txt
# streamlit==1.27.1
# requests==2.31.0
# numpy==1.26.0
# faiss-cpu==1.7.4
# fpdf2==2.7.6  # Note the new package name
# jinja2==3.1.2
# langchain==0.0.349
# python-dotenv==1.0.0  # Add if using environment variables


streamlit==1.16.0
requests==2.28.2
numpy==1.23.1
faiss-cpu==1.7.3
fpdf==1.7.2
jinja2==3.1.2
langchain
altair==4.2.0
